Some refactoring to MergePolicyFactory
(*) Remove MergePolicyFactoryHelper. MergePolicyFactory now takes a
SolrResourceLoader for creating MergePolicy and MergePolicyFactory
instances.

(*) Adds support for "wrapped.prefix" key for a
WrapperMergePolicyFactory. This is used to configure the wrapped
MergePolicyFactory.

(*) WrapperMergePolicyFactory defines an abstract
getDefaultMergePolicy() so implementations can return the default MP
in case none is configured.

(*) WrapperMergePolicyFactory only implements getting the wrapped
MergePolicy instance, while implementations return the actual final
MergePolicy.

(*) SimpleMergePolicyFactory requires implementations to return the MP
instance, while it then configures it with the set properties.

(*) Changed SolrIndexConfig to not fail (for now) if solrconfig.xml
contains <mergePolicy> but not the <mPFactory>, so that in this commit
we can focus on getting the actual classes in place. In a follow-on
commit we should remove support for <mergePolicy> in trunk and
deprecate in 5x.

(*) Added RandomMergePolicyFactory for tests.

(*) Removed all the added tests since the factory and associated
classes were moved to solr/ code.

(*) Added WrapperMergePolicyFactoryTest.