SimpleMergePolicyFactory is now no longer abstract. Shall we remove LogByteSizeMergePolicyFactory/LogDocMergePolicyFactory/TieredMergePolicyFactory now then?
illustrations and observations:

(1)

<!-- SimpleMergePolicyFactory allows users to configure their own simple merge policies _without_ needing to write a corresponding merge policy factory. -->
<mergePolicyFactory class="solr.SimpleMergePolicyFactory">
  <str name="class">com.mycompany.MySimpleMergePolicy</str> <!-- required element -->
  ...
</mergePolicyFactory>

(2)

(a) <!-- SimpleMergePolicyFactory however then also allows users to configure one of the out-of-the-box merge policies. -->
<mergePolicyFactory class="solr.SimpleMergePolicyFactory">
  <str name="class">solr.TieredMergePolicy</str> <!-- required element -->
  ...
</mergePolicyFactory>

(b) <!-- Since/If the out-of-the-box merge policy has a corresponding factory this would be equivalent to configuring the factory _without_ a class element. -->
<mergePolicyFactory class="solr.TieredMergePolicyFactory">
  ...
</mergePolicyFactory>

(c) Might there being alternatives (a) and (b) be confusing? Though perhaps (b) could remain undocumented?

(3)

(a) <!-- If the out-of-the-box merge policy has a corresponding factory then it could be customised with a class element. -->
<mergePolicyFactory class="solr.TieredMergePolicyFactory">
  <str name="class">com.mycompany.MyTieredMergePolicy</str>
  ...
</mergePolicyFactory>

(b) Is that enough 'added value' for TieredMergePolicyFactory to exists? I suspect not.

(4)

(a) For the wrapper policies we already have the idea of a "mergePolicy.class" element for the _wrapped_ policy and so extending that (optionally) to the _wrapping_ i.e. top-level class seems logical (at least to me) _in_the_context_ of SimpleMergePolicyFactory taking a "class" element:

<mergePolicyFactory class="UpgradeIndexMergePolicyFactory">
  <str name="class">com.mycompany.MyUpgradeIndexMergePolicy</str> <!-- optional element: not all WrapperMergePolicyFactory implementations require a "class" element -->
  <str name="wrapped.key">mergePolicy</str>
  <str              name="mergePolicy.class">TieredMergePolicyFactory</str>
  <double           name="mergePolicy.segmentsPerTier">4.2</double>
</mergePolicyFactory>

(b) Notice that we now have "class" meaning MergePolicy in the wrapper context and MergePolicyFactory in the wrapped context. How about disambiguating like this?

<mergePolicyFactory class="UpgradeIndexMergePolicyFactory">
  <str name="class">com.mycompany.MyUpgradeIndexMergePolicy</str> <!-- optional element: not all WrapperMergePolicyFactory implementations require a "class" element -->
  <str name="wrapped.key">mergePolicy</str>
  <str              name="mergePolicy.factoryClass">TieredMergePolicyFactory</str>
  <double           name="mergePolicy.segmentsPerTier">4.2</double>
</mergePolicyFactory>

(c) Also notice that if the TieredMergePolicyFactory were to be removed then the above example would actually become this:

<mergePolicyFactory class="UpgradeIndexMergePolicyFactory">
  <str name="class">com.mycompany.MyUpgradeIndexMergePolicy</str> <!-- optional element: not all WrapperMergePolicyFactory implementations require a "class" element -->
  <str name="wrapped.key">mergePolicy</str>
  <str              name="mergePolicy.factoryClass">SimpleMergePolicyFactory</str>
  <str              name="mergePolicy.class">solr.TieredMergePolicy</str>
  <double           name="mergePolicy.segmentsPerTier">4.2</double>
</mergePolicyFactory>

Or in other words if we wish to remove the TieredMergePolicyFactory _and_ have support for a "class" element for both SimpleMergePolicyFactory and WrapperMergePolicyFactory then WrapperMergePolicyFactory must use a "factoryClass" element for its wrapped policy because the wrapped policy itself will potentially need the "class" element.